+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v1.2.0
+* Googly eyes on creatures
 * Support for per-mod googly-eye configuration:
   Create a file googly-eye-config.json in the mod's resources (next to ModTheSpire.json).